% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.r
\name{getColumns}
\alias{getColumns}
\title{Lists the column names of a table and their types}
\usage{
getColumns(table_name, config = getDefaultConfig())
}
\description{
Lists the column names of a table and their types
}

